+++ src/Client/ClientCore/Rendering/Sprites/SpriteSheet.cs
@@ -16,6 +16,8 @@
  */
 
 using System;
+using System.Runtime.InteropServices;
+using Sovereign.ClientCore.Rendering.Configuration;
 
 namespace Sovereign.ClientCore.Rendering.Sprites;
 
@@ -33,7 +35,7 @@ public class SpriteSheet : IDisposable
     /// <summary>
     ///     Spritesheet definition.
     /// </summary>
-    public SpriteSheetDefinition Definition { get; private set; }
+    public SpriteSheetDefinition Definition { get; }
 
     /// <summary>
     ///     SDL_Surface holding the spriteset.
@@ -43,5 +45,42 @@ public class SpriteSheet : IDisposable
     public void Dispose()
     {
         Surface.Dispose();
+    }
+
+    /// <summary>
+    ///     Checks the opacity of the sprite at the given row and column.
+    /// </summary>
+    /// <param name="row">Row.</param>
+    /// <param name="column">Column.</param>
+    /// <returns>true if the sprite is completely opaque; false otherwise.</returns>
+    public bool CheckOpacity(int row, int column)
+    {
+        // We're going to test the opacity of the sprite by checking that every pixel has the
+        // maximum alpha value. If every pixel has max alpha, then there is no transparency and
+        // the renderer will not blend the sprite with the pixels beneath it; therefore, the
+        // sprite is opaque.
+
+        // Grab the raw pixel data for the sprite. Spritesheets are converted to the display format
+        // when loaded, so we don't have to worry about conversions here. We're assuming 4 bytes per
+        // pixel as these are the only display formats supported by the engine.
+        var sheetPixelData = Surface.Properties.Data;
+        var alphaMask = Surface.Properties.Format switch
+        {
+            DisplayFormat.B8G8R8A8_UNorm => 0xFF000000,
+            DisplayFormat.R8G8B8A8_UNorm => 0xFF000000,
+            _ => 0xFF000000
+        };
+
+        for (var x = column * Definition.SpriteWidth; x < (column + 1) * Definition.SpriteWidth; ++x)
+        {
+            for (var y = row * Definition.SpriteHeight; y < (row + 1) * Definition.SpriteHeight; ++y)
+            {
+                var byteOffset = y * Surface.Properties.Pitch + x * Surface.Properties.BytesPerPixel;
+                var pixelValue = (uint)Marshal.ReadInt32(sheetPixelData, (int)byteOffset);
+                if ((pixelValue & alphaMask) != alphaMask) return false;
+            }
+        }
+
+        return true;
     }
 }